Bug fixes. Added PetListUI with controller and service. Fixed error messages. Added AbstractClickHandler.

+++ src/main/java/com/netcracker/service/petprofile/PetProfileService.java
@@ -2,6 +2,8 @@ package com.netcracker.service.petprofile;
 
 import com.netcracker.model.pet.Pet;
 import com.netcracker.model.user.Profile;
+
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
@@ -16,7 +18,7 @@ public interface PetProfileService {
     void editProfile(Profile profile);
 
 //    Система должна позволять владельцу питомца удалять страницу питомца.
-    void deleteProfile(Integer profileId);
+    void deleteProfile(BigInteger profileId);
 
 //    Система должна позволять владельцу питомца изменить хозяина питомца (при продаже/передачи питомца по объявлению
     void changePetOwner(Profile owner, Profile newOwner);
@@ -24,6 +26,6 @@ public interface PetProfileService {
     Date ageCalculation(Pet pet);
 
     //Получить список животных переданного профиля
-    List<Pet> getAllProfilePets(Integer profileId);
+    List<Pet> getAllProfilePets(BigInteger profileId);
 
 }
